Refactoring graphics module a little.
There's gonna be piles of trait impl's for conversions on the Color
type, and probably others, so I'm pulling those out into their own
little submodule.
